rename "calc" module to "data_shapshot"
