+++ packages/python/genai_prices/data_snapshot.py
