+++ tests/test_custom_prices.py
@@ -8,7 +8,7 @@ import pytest
 from inline_snapshot import snapshot
 
 from genai_prices import Usage, calc_price, data, types
-from genai_prices.calc import DataSnapshot, set_custom_snapshot
+from genai_prices.data_snapshot import DataSnapshot, set_custom_snapshot
 from genai_prices.update_prices import UpdatePrices
