+++ tests/test_update_prices.py
@@ -4,7 +4,14 @@ import httpx
 import pytest
 from inline_snapshot import snapshot
 
-from genai_prices import UpdatePrices, Usage, calc, calc_price, wait_prices_updated_async, wait_prices_updated_sync
+from genai_prices import (
+    UpdatePrices,
+    Usage,
+    calc_price,
+    data_snapshot,
+    wait_prices_updated_async,
+    wait_prices_updated_sync,
+)
 
 pytestmark = pytest.mark.anyio
 
@@ -12,11 +19,11 @@ pytestmark = pytest.mark.anyio
 @pytest.mark.default_cassette('success.yaml')
 @pytest.mark.vcr()
 def test_update_prices_wait_on_start():
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
     with UpdatePrices() as update_prices:
-        assert calc._custom_snapshot is None
+        assert data_snapshot._custom_snapshot is None
         update_prices.wait()
-        assert calc._custom_snapshot is not None
+        assert data_snapshot._custom_snapshot is not None
         price = calc_price(Usage(input_tokens=1000, output_tokens=100), model_ref='gpt-4o', provider_id='openai')
         assert price.input_price == snapshot(Decimal('0.0025'))
         assert price.output_price == snapshot(Decimal('0.001'))
@@ -28,11 +35,11 @@ def test_update_prices_wait_on_start():
 @pytest.mark.default_cassette('success.yaml')
 @pytest.mark.vcr()
 def test_wait_prices_updated_sync():
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
     with UpdatePrices():
-        assert calc._custom_snapshot is None
+        assert data_snapshot._custom_snapshot is None
         wait_prices_updated_sync()
-        assert calc._custom_snapshot is not None
+        assert data_snapshot._custom_snapshot is not None
         price = calc_price(Usage(input_tokens=1000, output_tokens=100), model_ref='gpt-4o', provider_id='openai')
         assert price.input_price == snapshot(Decimal('0.0025'))
         assert price.output_price == snapshot(Decimal('0.001'))
@@ -44,11 +51,11 @@ def test_wait_prices_updated_sync():
 @pytest.mark.default_cassette('success.yaml')
 @pytest.mark.vcr()
 async def test_wait_prices_updated_async():
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
     with UpdatePrices():
-        assert calc._custom_snapshot is None
+        assert data_snapshot._custom_snapshot is None
         await wait_prices_updated_async()
-        assert calc._custom_snapshot is not None
+        assert data_snapshot._custom_snapshot is not None
         price = calc_price(Usage(input_tokens=1000, output_tokens=100), model_ref='gpt-4o', provider_id='openai')
         assert price.input_price == snapshot(Decimal('0.0025'))
         assert price.output_price == snapshot(Decimal('0.001'))
@@ -60,22 +67,22 @@ async def test_wait_prices_updated_async():
 @pytest.mark.default_cassette('fail.yaml')
 @pytest.mark.vcr()
 def test_update_prices_failed():
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
     with UpdatePrices(url='https://demo-endpoints.pydantic.workers.dev/bin?status=404') as update_prices:
         with pytest.raises(httpx.HTTPStatusError):
             update_prices.wait()
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
 
 
 @pytest.mark.default_cassette('fail.yaml')
 @pytest.mark.vcr()
 def test_update_prices_failed_stop():
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
     update_prices = UpdatePrices(url='https://demo-endpoints.pydantic.workers.dev/bin?status=404')
     update_prices.start()
     with pytest.raises(httpx.HTTPStatusError):
         update_prices.stop()
-    assert calc._custom_snapshot is None
+    assert data_snapshot._custom_snapshot is None
 
 
 @pytest.mark.default_cassette('success.yaml')
